+++ tests/utils/theme/ThemeProvider.test.tsx
@@ -22,6 +22,7 @@ function setupMatchMedia(matches: boolean) {
         })),
     });
 }
+
 describe("ThemeProvider", () => {
     afterEach(() => {
         cleanup();
